+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Possible log types:
 - `[fixed]` for any bug fixes.
 - `[security]` to invite users to upgrade in case of vulnerabilities.
 
-### v0.4.0 (2025-04-29) - (unreleased)
+### v0.4.0 (2025-05-30) - (unreleased)
 
 - [added] Support for macro expressions.  See `MacroDecimal`.
 - [added] `MacroBoolean`.
@@ -19,6 +19,7 @@ Possible log types:
 - [changed] Replaced `Aperture::Other` with `Aperture::Macro`, the latter having option arguments.
 - [changed] Use `MacroBoolean` instead of `bool` for `exposure` on macro primitives.
 - [changed] `PolygonPrimitive::new` uses a `MacroInteger` for the `vertices` argument.
+- [changed] Replaced dependency on `conv::TryFrom` with `std::convert::TryFrom` (since TryFrom was stabilized in Rust 1.34).
 - [deprecated] `*Primitive::exposure_on`, replaced with `*Primitive::with_exposure`.
 
 ### v0.3.0 (2022-07-05)